[amp] refactor to consolidate repeated code
